backend/fix: update types for LTS nearby drivers api

+++ Backend/app/rider-platform/rider-app/Main/src/SharedLogic/External/LocationTrackingService/Types.hs
@@ -108,7 +108,7 @@ data NearByDriverRes = NearByDriverRes
     coordinatesCalculatedAt :: UTCTime,
     createdAt :: UTCTime,
     updatedAt :: UTCTime,
-    bear :: Maybe Int,
+    bear :: Maybe Double,
     vehicleType :: VV.VehicleVariant,
     rideDetails :: Maybe RideDetails
   }
